Added script for generating the Admin account. Enforced linting on amdin scripts.

+++ admin/purge-database.js
@@ -1,11 +1,11 @@
 /****************************************************************************************
- * 
+ *
  * USAGE:
  * gulp purge-database [mongoDbEndpoint]
- * 
+ *
  * If not supplied on the command line, mongoDbEndpoint will default to
  * mongodb://localhost/dev
- * 
+ *
  ***************************************************************************************/
 
 import config from '../service/config';
@@ -17,6 +17,7 @@ if (process.argv[3]) {
 import database from '../service/data/database';
 import log from 'fancy-log';
 import LogEntry from '../service/data/log-entry';
+import readline from 'readline-sync';
 import Session from '../service/data/session';
 import User from '../service/data/user';
 
@@ -24,11 +25,19 @@ module.exports = async function () {
 	log(`Using MongoDB endpoint "${ config.mongoEndpoint }"...`);
 
 	try {
+		if (!readline.keyInYN(
+			'This operation will totally destroy all data in the database. Are you sure you want to proceed?'
+		)) {
+			log('Exiting.');
+			database.connection.close();
+			return;
+		}
+
 		log('Purging DB tables...');
 		await Session.deleteMany({});
 		await LogEntry.deleteMany({});
 		await User.deleteMany({});
-		log('Done')
+		log('Done.');
 	} catch (err) {
 		log.error(err);
 		process.exitCode = 1;
